requirements.txt
APScheduler==3.8.0
clipboard==0.0.4
Pillow==8.3.2
pyperclip==1.8.2
pytz==2021.1
six==1.16.0
tzlocal==2.1
